// src/generated/api/model/organisationMemberGetListParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NestJS Boilerplate API
 * API documentation for the NestJS boilerplate project.
 * OpenAPI spec version: 1.0
 */
import type { Object } from './object';

export type OrganisationMemberGetListParams = {
page: Object;
limit: Object;
};